Add extra sanity check for commit showing up multiple times
This happened in the wild when someone pushed some commits to a remote
branch, did a local edit and rebase, and then re-rebased on top of the
old commits.  Git will end up with two "copies" of the commit; one
with the base changes, and the other with the delta (because it
knows how to resolve conflicts when both commits did the same
modification).  Be nice and give a good error in this case.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>