Improve ghstack land to "merge" and delete branches. (#46)
I realized there is a simple trick for convincing GitHub that a PR
has been merged, rather than closed: merge the head branch into the
base branch!  We can't (easily) induce this when we do an actual master
merge commit, because there isn't actually any relationship between the
master commit and the pull request (which is a request to merge head
into base--master is completely unrelated).

Also, because people really seem to hate it when ghstack leaves tons
of branches lying around, I also updated the logic to delete the
branches at merge time (this doesn't lose data, because GitHub keeps
track of the pull request branch at time of merge itself).  One
slightly goofy thing is the deletion of branches will cause ghstack
to reuse GhNumbers, but as far as I can tell this is completely
harmless.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

[ghstack-poisoned]