Undo Patch abstraction
I am trying to implement incremental updates and the patch abstraction
is the wrong way to go about doing it, because you in general don't have
a base tree to apply the patch on; you need the full git tree object
somehow.  This will break ghexport when we take these updates but it's
going to be complicated to handle anyway, deal with it when we get
there.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: 3e8841044f95af9f83452ec0daf482dc8dbc12f5
Pull Request resolved: https://github.com/ezyang/ghstack/pull/198