Make Pull-Request-resolved hyphenated to be valid trailer
Previously, we used the syntax "Pull Request resolved", however,
this is not a valid Git trailer as there are spaces in the key.
Replacing the spaces with hyphens fixes this, while not impeding
GitHub's auto-resolve capabilities (as it only matches for resolved
at the tail of the string).

I also use git interpret-trailers to handle trailer modification,
which makes ghstack more robust when there are pre-existing trailers
in the commit message (as all trailers must be together in a blog,
there must not be any double newlines.)

Signed-off-by: Edward Z. Yang <ezyang@mit.edu>

ghstack-source-id: dfa949d56854ddc1448e9108920b95c6b2de1e65
ghstack-comment-id: 2726090606
Pull Request resolved: https://github.com/ezyang/ghstack/pull/277